fixed regurgitator hungry persisting with no mana fixed vortex ritual persisting with no mana and not having patreon noted in tooltip ice queen crown buff: slows nearby enemies and projectiles when you take damage (with toggle)

+++ Patreon/GreatestKraken/VortexRitualProj.cs
@@ -82,8 +82,11 @@ namespace FargowiltasSouls.Patreon.GreatestKraken
             //kill me if player is not holding
             Player player = Main.player[projectile.owner];
 
-            if (player.dead || !player.active || !(player.HeldItem.type == ModContent.ItemType<VortexMagnetRitual>() && player.channel))
+            if (player.dead || !player.active || !(player.HeldItem.type == ModContent.ItemType<VortexMagnetRitual>() && player.channel && player.CheckMana(player.HeldItem.mana)))
+            {
                 projectile.Kill();
+                return;
+            }
 
             projectile.damage = player.GetWeaponDamage(player.HeldItem);
             projectile.knockBack = player.GetWeaponKnockback(player.HeldItem, player.HeldItem.knockBack);
